fix: Viseme Float型修正, AAO互換性, ビルド後コンポーネント破棄
- VisemeAnimatorGenerator: BlendTreeパラメータ型をInt→Floatに修正
- PiercingPass: ビルド後にPiercingSetupをDestroyImmediate
- AAO API連携: ComponentInformation<PiercingSetup>を登録し未知コンポーネント警告を解消
- asmdef: com.anatawa12.avatar-optimizer.api.editor参照とPIERCING_AAO定義を追加

+++ Editor/Processing/VisemeAnimatorGenerator.cs
@@ -96,7 +96,7 @@ namespace PiercingTool.Editor
             controller.name = "PiercingViseme";
 
             // Visemeパラメータ追加
-            controller.AddParameter("Viseme", AnimatorControllerParameterType.Int);
+            controller.AddParameter("Viseme", AnimatorControllerParameterType.Float);
 
             // レイヤー追加
             controller.AddLayer("PiercingViseme");
